% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_titles.R
\name{clean_titles}
\alias{clean_titles}
\title{Clean up column names in a dataframe}
\usage{
clean_titles(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe with cleaned up column names, in snake case.
}
\description{
Converts dataframe column names to snake case.
}
\examples{

clean_titles(data = iris)
}
